GUI Move control helper utils to own module
